MonsterHealth 컴포넌트 제거 후 Monster 컴포넌트에 통합
