--- Unity/Assets/Monster/MonsterHealth.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 8337cfdb1a4542448ef4b10c117271cb
-timeCreated: 1701047188
